Deduplicate lifecycle PHP snippets
