+++ packages/runtime-core/src/index.ts
@@ -13,6 +13,7 @@ export * from "./runtime-command-result.js"
 export * from "./runtime-backend-resolver.js"
 export * from "./runtime-defaults.js"
 export * from "./runtime-env.js"
+export * from "./runtime-php-snippets.js"
 export * from "./runtime-overlay-bundle.js"
 export * from "./redaction.js"
 export * from "./runtime-policy.js"
